RacingController 클래스 추가 및 domain, view 패키지 분리
