+++ src/main/kotlin/racingcar/Main.kt
@@ -1,11 +1,7 @@
 package racingcar
 
+import racingcar.view.RacingController
+
 fun main () {
-    val inputData: InputData = InputView().view()
-
-    val cars: List<Car> = Cars(inputData.inputCarNames).cars
-    val racing = Racing(cars = cars, inputData.inputAttemptCount)
-
-    val zeroToNineRandomIndex: () -> Int = {(0..9).random()}
-    ResultView().view(racing.start(carForwardRandomProvider = zeroToNineRandomIndex))
+    RacingController().start()
 }
