+++ src/main/kotlin/racingcar/view/InputData.kt
@@ -1,3 +1,3 @@
-package racingcar
+package racingcar.view
 
 data class InputData(val inputCarNames: String, val inputAttemptCount: Int)
