+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.view
 
 class InputView {
     private fun inputCarNames(): String {
