+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.view
+
+import racingcar.domain.RacingResult
 
 
 class ResultView {
