+++ src/test/kotlin/racingcar/CarForwardAttemptTest.kt
@@ -6,6 +6,9 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.domain.Car
+import racingcar.domain.CarForwardAttempt
+import racingcar.domain.Racing
 import java.lang.IllegalArgumentException
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
